+++ test/unit/test_feat/test_feat_fe.c
@@ -98,7 +98,7 @@ main(int argc, char *argv[])
 	for (i = 0; i < total_frames; ++i) {
 		int32 j;
 		printf("%-4d ", i);
-		for (j = 0; j < feat_dimension(fcb); ++j) {
+		for (j = 0; (uint32)j < feat_dimension(fcb); ++j) {
 			TEST_EQUAL_FLOAT(featbuf1[i][0][j], featbuf2[i][0][j]);
 		}
 		if (i % 10 == 9)
@@ -134,7 +134,7 @@ main(int argc, char *argv[])
 	for (i = 0; i < total_frames; ++i) {
 		int32 j;
 		printf("%-4d ", i);
-		for (j = 0; j < feat_dimension(fcb); ++j)
+		for (j = 0; (uint32)j < feat_dimension(fcb); ++j)
 			TEST_EQUAL_FLOAT(featbuf1[i][0][j], featbuf2[i][0][j]);
 		if (i % 10 == 9)
 			printf("\n");
